+++ src/main/kotlin/dev/aaronhowser/mods/aaron/client/AaronClientUtil.kt
@@ -1,8 +1,12 @@
 package dev.aaronhowser.mods.aaron.client
 
 import net.minecraft.client.Minecraft
+import net.minecraft.client.resources.language.I18n
+import net.minecraft.network.chat.Component
+import net.minecraft.resources.ResourceKey
 import net.minecraft.world.entity.player.Player
 import net.minecraft.world.level.Level
+import net.minecraft.world.level.biome.Biome
 import net.neoforged.fml.loading.FMLEnvironment
 
 object AaronClientUtil {
@@ -20,5 +24,16 @@ object AaronClientUtil {
 			if (!FMLEnvironment.dist.isClient) return null
 			return Minecraft.getInstance().player
 		}
+
+	fun getBiomeDisplay(biomeKey: ResourceKey<Biome>): Component {
+		val probableTranslationKey = "biome.${biomeKey.location().namespace}.${biomeKey.location().path}"
+		val hasTranslation = I18n.exists(probableTranslationKey)
+
+		return if (hasTranslation) {
+			Component.translatable(probableTranslationKey)
+		} else {
+			Component.literal(biomeKey.location().toString())
+		}
+	}
 
 }